+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fixed value order in world files.
 * Added more paper formats. [#7](https://github.com/Zverik/Nik4/issues/7)
+* Style XML can now be streamed from stdin.
 
 ## 1.2, 19.05.2014